refactor(benchmarks): route scenes through workload controllers
